moved components not used
