+++ burp_reports/backends/notusedyet/test.py
@@ -1,5 +1,4 @@
-from burpui.misc.backend.burp2 import Burp
-from burp_reports.backends.burpui_clients import Clients
+from burp_reports.backends.notusedyet.burpui_clients import Clients
 clientes = Clients(burp_version=2, conf='/etc/burp/burp-ui.cfg')
 
 clientes.get_clients()
